+++ CHANGELOG.md
@@ -16,6 +16,11 @@ Types of changes:
 ### Fixed
 
 - Germline risk VCF is now a temp file (#303)
+- Include flag now works for metrics report (#304)
+
+### Added
+
+- Additional test cases for metrics report (#304)
 
 ## [3.1.2] - 2026-02-02